+++ examples/src/features/module_nesting.rs
@@ -1,8 +1,6 @@
-use odra::{
-    contract_env::emit_event,
-    prelude::{string::String, vec::Vec},
-    Event, List, Mapping, OdraType, UnwrapOrRevert, Variable
-};
+use casper_event_standard::Event;
+use odra::prelude::*;
+use odra::{Mapping, Module, ModuleWrapper, OdraType, UnwrapOrRevert, Variable};
 
 #[odra::module]
 pub struct ResultsStorage {
@@ -13,8 +11,7 @@ pub struct ResultsStorage {
 #[odra::module]
 pub struct NestedOdraTypesContract {
     latest_result: Variable<OperationResult>,
-    current_generation_storage: ResultsStorage,
-    past_generations_storage: List<ResultsStorage>
+    current_generation_storage: ModuleWrapper<ResultsStorage>
 }
 
 #[odra::module]
@@ -36,31 +33,11 @@ impl NestedOdraTypesContract {
             .results_count
             .set(results_count);
 
-        emit_event(OperationEnded {
-            result: operation_result
+        self.env().emit_event(OperationEnded {
+            id: operation_result.id,
+            status: operation_result.status,
+            description: operation_result.description
         });
-    }
-
-    pub fn new_generation(&mut self) {
-        // move current generation to past generations
-        let keys = self
-            .current_generation_storage
-            .results_count
-            .get_or_default();
-        let keys_range = 0..keys;
-        let mut past_generation = self.past_generations_storage.next_instance();
-
-        keys_range.for_each(|key| {
-            let current_result = self
-                .current_generation_storage
-                .results
-                .get(&key)
-                .unwrap_or_revert();
-            past_generation.results.set(&key, current_result);
-            past_generation.results_count.set(key + 1);
-        });
-
-        self.current_generation_storage.results_count.set(0);
     }
 
     pub fn latest_result(&self) -> Option<OperationResult> {
@@ -78,22 +55,7 @@ impl NestedOdraTypesContract {
                 self.current_generation_storage
                     .results
                     .get(&key)
-                    .unwrap_or_revert()
-            })
-            .collect()
-    }
-
-    pub fn past_generations(&mut self) -> Vec<Vec<OperationResult>> {
-        let keys = self.past_generations_storage.len();
-        let keys_range = 0..keys;
-        keys_range
-            .map(|key| {
-                let generation = self.past_generations_storage.get_instance(key);
-                let keys = generation.results_count.get_or_default();
-                let keys_range = 0..keys;
-                keys_range
-                    .map(|key| generation.results.get(&key).unwrap_or_revert())
-                    .collect()
+                    .unwrap_or_revert(&self.env())
             })
             .collect()
     }
@@ -114,16 +76,14 @@ pub struct OperationResult {
 
 #[derive(Event, PartialEq, Debug)]
 pub struct OperationEnded {
-    result: OperationResult
+    id: u32,
+    status: Status,
+    description: String
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use odra::{
-        assert_events,
-        prelude::{string::ToString, vec}
-    };
 
     // generate operation results
     fn operation_results() -> Vec<OperationResult> {
@@ -173,7 +133,8 @@ mod tests {
 
     #[test]
     fn nested_odra_types() {
-        let mut nested_odra_types = NestedOdraTypesContractDeployer::default();
+        let test_env = odra::test_env();
+        let mut nested_odra_types = NestedOdraTypesContractDeployer::init(&test_env);
 
         // Storage is not set
         assert_eq!(nested_odra_types.latest_result(), None);
@@ -188,62 +149,13 @@ mod tests {
         );
 
         // Events are also saved properly, even with nested Odra Types
-        assert_events!(
-            nested_odra_types,
-            OperationEnded {
-                result: operation_results()[0].clone()
+        test_env.emitted_event(
+            &nested_odra_types.address,
+            &OperationEnded {
+                id: operation_results()[0].clone().id,
+                status: operation_results()[0].clone().status,
+                description: operation_results()[0].clone().description
             }
-        );
-    }
-
-    #[test]
-    fn module_in_mapping_in_mapping() {
-        let mut nested_odra_types = NestedOdraTypesContractDeployer::default();
-        let operations = operation_results();
-        // fill the storage
-        nested_odra_types.save_operation_result(operations[0].clone());
-        nested_odra_types.save_operation_result(operations[1].clone());
-        nested_odra_types.save_operation_result(operations[2].clone());
-
-        // create a new generation
-        nested_odra_types.new_generation();
-
-        // fill the storage
-        nested_odra_types.save_operation_result(operations[3].clone());
-        nested_odra_types.save_operation_result(operations[4].clone());
-        nested_odra_types.save_operation_result(operations[5].clone());
-
-        // create a new generation
-        nested_odra_types.new_generation();
-
-        // fill the storage last time
-        nested_odra_types.save_operation_result(operations[6].clone());
-        nested_odra_types.save_operation_result(operations[7].clone());
-
-        // check the storage
-        assert_eq!(
-            nested_odra_types.latest_result(),
-            Some(operations[7].clone())
-        );
-        assert_eq!(
-            nested_odra_types.current_generation(),
-            vec![operations[6].clone(), operations[7].clone(),]
-        );
-
-        assert_eq!(
-            nested_odra_types.past_generations(),
-            vec![
-                vec![
-                    operations[0].clone(),
-                    operations[1].clone(),
-                    operations[2].clone()
-                ],
-                vec![
-                    operations[3].clone(),
-                    operations[4].clone(),
-                    operations[5].clone()
-                ]
-            ]
         );
     }
 }